Move JS in recipe-app-code folder to docs main code file and delete recipe-app-code foilder
